[ci skip] More use/project hierarchy refactoring
